requirements.txt
python=3.7
dxchange=0.1.8
matplotlib=3.5.1
pyqtgraph=0.11
seaborn=0.11.2
setuptools
pyqt =5.9.2
pywin32 # [win]
tomopy=1.11.0
scikit-image=0.19.2
numpy=1.21.5